Basic Main Menu and Score
Added a small and basic Main Menu (It looks bad) that asks for the player's name and records it to playerPrefs in order to keep it inbetween scenes. Also applied score giving to the minigames for the most part, I think. I may have missed a part of the script where its supposed to give you points if so please inform me.